pyproject.toml
[tool.poetry]
name = "crawler-python"
version = "0.1.0"
description = "Crawler em Python com Selenium, Beautiful Soup e Pandas"
authors = ["Lemes"]
readme = "README.md"
packages = [{ include = "crawler", from = "." }]

[tool.poetry.dependencies]
python = "^3.11"
selenium = "^4.27"
beautifulsoup4 = "^4.12"
pandas = "^2.2"
python-dotenv = "^1.0"
tqdm = "^4.65"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
